+++ src/components/App/index.js
@@ -1,11 +1,9 @@
 import React from 'react';
-import logo from './logo.svg';
-import './App.css';
+import './style.css';
 
 const App = () => (
   <div className="App">
     <header className="App-header">
-      <img src={logo} className="App-logo" alt="logo" />
       <h1 className="App-title">Hello, jovem!</h1>
     </header>
   </div>
